fix: use VM_PORT_ANY when binding vsock listeners and support u32 ports in URIs

+++ packages/uri/src/lib.rs
@@ -169,7 +169,7 @@ impl Uri {
 	}
 
 	#[must_use]
-	pub fn port(&self) -> Option<u16> {
+	pub fn port(&self) -> Option<u32> {
 		self.authority_raw().and_then(|authority| {
 			let authority = authority.split('@').next_back()?;
 			if authority.starts_with('[') {
@@ -181,7 +181,7 @@ impl Uri {
 	}
 
 	#[must_use]
-	pub fn port_or_known_default(&self) -> Option<u16> {
+	pub fn port_or_known_default(&self) -> Option<u32> {
 		self.port().or_else(|| match self.scheme()? {
 			"http" | "http+unix" => Some(80),
 			"https" => Some(443),
